+++ domain/index.ts
@@ -5,6 +5,8 @@ export * from './navigation';
 export * from './customer';
 export * from './order';
 export * from './membership';
+export * from './look';
+